% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.r
\name{reformatVoterList}
\alias{reformatVoterList}
\title{Reformat a list of new voters. For internal use during the 2024 April TRU-UE contract ratification vote.}
\usage{
reformatVoterList(filename_in, filename_out)
}
\description{
Reformat a list of new voters. For internal use during the 2024 April TRU-UE contract ratification vote.
}
